% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{compare_genotypes}
\alias{compare_genotypes}
\title{Compare microCT data between two genotypes}
\usage{
compare_genotypes(data, type = NULL)
}
\arguments{
\item{data}{Trabecular bone microCT data in a data frame, formatted as is the
output of \code{\link[=read_trabecular_csv]{read_trabecular_csv()}}, \code{\link[=read_cortical_csv]{read_cortical_csv()}}, or
\code{\link[=read_mfe_csv]{read_mfe_csv()}}. Note that data for only one Site can be supplied at a
time.}

\item{type}{A string indicating the type of data supplied. Options include
\code{trabecular}, \code{cortical}, or \code{fea}. Defaults to \code{NULL}, in which case the
function will try to figure out which type of data it is.}
}
\value{
A list of each bone measure analyzed. Each measure is itself a list
of each sex analyzed. Each sex is a tibble/data frame containing columns
for \code{Sex}, \code{Genotype}, \code{n}, \code{Mean}, \code{SEM}, \code{P}, and \code{Sig} (a string
containing \code{"*"} if \emph{P} < 0.05, \code{"**"} if \emph{P} < 0.01, or \code{"***"} if
\emph{P} < 0.001).
}
\description{
This function takes in trabecular, cortical, or finite element analysis bone
microCT data at one site and compares it between two genotypes using a
Student's \emph{t}-test.
}
\details{
If \code{trabecular} bone data are supplied, the measures compared include:
\itemize{
\item \code{BV/TV}
\item \code{SMI}
\item \code{Tb.N}
\item \code{Tb.Th}
\item \code{Tb.Sp}
}

If \code{cortical} bone data are supplied, the measures compared include:
\itemize{
\item \code{Ct.vBMD}
\item \code{Ct.Th}
\item \code{End.Circ}
\item \code{Peri.Circ}
\item \code{Ct.Po}
\item \code{Ct.Po.V}
}

If finite element analysis (\code{fea}) data are supplied, the measures compared
include:
\itemize{
\item \code{S}
\item \code{F.ult}
}
}
\examples{
gen_key <- read_key_csv(mctr_ex("example-gen-key.csv"))
gen_trab <- read_trabecular_csv(mctr_ex("example-trabecular.csv"),
                            gen_key)
Spine.Tb <- gen_trab |> dplyr::filter(Site == "Spine") |>
                compare_genotypes()
}
