% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_genotype_comparison}
\alias{create_genotype_comparison}
\title{Create a project skeleton to compare bone microCT data between genotypes}
\usage{
create_genotype_comparison(
  file_name = paste0(format(Sys.Date(), "\%F"), "-microCT-Genotype-Comparison.Rmd")
)
}
\arguments{
\item{file_name}{A string to name the project analysis R Markdown document.
Defaults to a file in the style of
\code{"YYYY-MM-DD-microCT-Genotype-Comparison.Rmd"}, where \code{YYYY-MM-DD}
represents today's date.}
}
\value{
Returns \code{TRUE} if successful.
}
\description{
This function assumes you are in a project root directory and creates an R
Markdown file containing a pre-filled bone microCT analysis comparing two
genotypes. It also creates a \code{data} directory containing template files into
which to paste the microCT data. Additionally, the \code{data} directory contains
a template \code{key.csv} file to specify the appropriate sample information.
}
\details{
The \code{data} directory contains the following files:
\itemize{
\item \code{key.csv}
\item \code{mfe.csv}
\item \code{trabecular.csv}
\item \code{twice1.csv}
\item \code{twice2.csv}
}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
create_genotype_comparison()
\dontshow{\}) # examplesIf}
}
