% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_organized_excel}
\alias{create_organized_excel}
\title{Write organized data to an Excel file}
\usage{
create_organized_excel(trab, cort, mfe, file = "Organized Data.xlsx")
}
\arguments{
\item{trab}{Trabecular microCT data, as formatted by the
\code{\link[=read_trabecular_excel]{read_trabecular_excel()}} function.}

\item{cort}{Cortical microCT data, as formatted by the
\code{\link[=read_cortical_excel]{read_cortical_excel()}} function.}

\item{mfe}{FEA microCT data, as formatted by the \code{\link[=read_mfe_excel]{read_mfe_excel()}} function.}

\item{file}{The name of the Excel file to write to. Defaults to \verb{Organized Data.xlsx}.}
}
\description{
This function takes microCT data which has been read in and had measures
calculated (as by the \verb{read_excel*()} functions), sorts it, and writes it to
an Excel file. Trabecular, cortical, and FEA data are reqired, and each is
written to its own sheet in the final Excel file.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
gen_key <- read_key_csv(mctr_ex("example-gen-key.csv"))
gen_trab <- read_trabecular_csv(mctr_ex("example-trabecular.csv"), gen_key)
gen_cort <- read_cortical_csv(mctr_ex("example-twice1.csv"),
                              mctr_ex("example-twice2.csv"), gen_key)
gen_mfe <- read_mfe_csv(mctr_ex("example-mfe.csv"), gen_key)
create_organized_excel(gen_trab, gen_cort, gen_mfe)
\dontshow{\}) # examplesIf}
}
