% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{excel_export_addin}
\alias{excel_export_addin}
\title{Convert a raw data Excel file to an organized Excel file}
\usage{
excel_export_addin()
}
\description{
This function is meant to be used as an RStudio addin. It uses \code{rstudioapi}
to ask the user to select an original/raw data Excel file, and then it asks
the user to create a file to save the organized Excel file.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
excel_export_addin()
\dontshow{\}) # examplesIf}
}
