% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{knit_with_colored_text}
\alias{knit_with_colored_text}
\title{Render an R Markdown document with colored text}
\usage{
knit_with_colored_text(input, ...)
}
\arguments{
\item{input}{The input file to be rendered.}

\item{...}{Additional arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
As for \code{\link[rmarkdown:render]{rmarkdown::render()}}, when \code{run_pandoc = TRUE}, the compiled
document is written into the output file, and the path of the output file
is returned. When \code{run_pandoc = FALSE}, the path of the Markdown output
file, with attributes \code{knit_meta} (the knitr meta data collected from code
chunks) and \code{intermediates} (the intermediate files/directories generated
by \code{\link[rmarkdown:render]{rmarkdown::render()}}).
}
\description{
This function is a wrapper around \code{\link[rmarkdown:render]{rmarkdown::render()}} which supplies a
custom Lua Pandoc filter to color text.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
knit_with_colored_text("example_file.Rmd")
\dontshow{\}) # examplesIf}
}
