% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_groups}
\alias{plot_groups}
\title{Plot bone microCT data by genotype or treatment}
\usage{
plot_groups(data, type = NULL, title = "sex", x_axis_angle = 45)
}
\arguments{
\item{data}{A tibble/data frame containing bone microCT data, formatted as is
the output of \code{\link[=read_trabecular_csv]{read_trabecular_csv()}}, \code{\link[=read_cortical_csv]{read_cortical_csv()}}, or
\code{\link[=read_mfe_csv]{read_mfe_csv()}} (or \verb{read_*_excel()}).}

\item{type}{A string indicating the type of data supplied. Options include
\code{trabecular}, \code{cortical}, or \code{fea}. Alternatively, a specific microCT
measure of interest can be specified (e.g., \code{Ct.vBMD}). Defaults to \code{NULL},
in which case the function will try to figure out which type of data it is.}

\item{title}{A string indicating what type of title the plots should have.
Defaults to \code{sex}, which is currently the only option implemented. To
remove titles from the plots, set \code{title} to \code{NULL}.}

\item{x_axis_angle}{The angle at which to rotate the x-axis labels so they
don't overlap, passed on to \code{\link[ggplot2:theme]{ggplot2::theme()}} as \code{axis.text.x = element_text(angle = x_axis_angle)}.}
}
\value{
A list containing \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} objects for sex each analyzed.
To print each plot without printing the list index, see \code{\link[=print_plots]{print_plots()}}.
}
\description{
This function takes in bone microCT data and produces boxplots comparing two
genotypes or treatments.
}
\details{
If the two genotypes or treatments are significantly different, statistical
significance will be indicated with \code{"*"} if \emph{P} < 0.05, \code{"**"} if \emph{P} <
0.01, or \code{"***"} if \emph{P} < 0.001.
}
\examples{
gen_key <- read_key_csv(mctr_ex("example-gen-key.csv"))
gen_cort <- read_cortical_csv(mctr_ex("example-twice1.csv"),
                          mctr_ex("example-twice2.csv"),
                          gen_key)
plot_groups(gen_cort |> dplyr::filter(Site == "Dia"))
}
