% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{print_plots}
\alias{print_plots}
\title{Print a list of plots}
\usage{
print_plots(plots)
}
\arguments{
\item{plots}{A list of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} objects.}
}
\value{
Returns NULL, since this function is used for its side effects.
}
\description{
This function iterates through a list of plots and prints them (without
printing the list index).
}
\examples{
key <- read_key_csv(microCTr_example("example-key.csv"))
cort <- read_cortical_csv(microCTr_example("example-twice1.csv"),
                          microCTr_example("example-twice2.csv"),
                          key)
plot_genotypes(cort |> dplyr::filter(Site == "Dia")) |> print_plots()
}
