% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_results}
\alias{print_results}
\title{Print bone mciroCT comparison analysis results}
\usage{
print_results(results, ...)
}
\arguments{
\item{results}{A list of microCT comparison results, formatted as is the
output of \code{\link[=compare_genotypes]{compare_genotypes()}} or \code{\link[=compare_treatments]{compare_treatments()}}.}

\item{...}{Additional arguments passed on to \code{\link[knitr:kable]{knitr::kable()}}. This function
uses a local default of \code{digits = 3} to specify how many significant digits
to print. This can be modified by passing a user specified \code{digits} value.}
}
\value{
Text output which is by default a Pandoc markdown pipe table for each
measure and each sex analyzed. If two sexes are analyzed, Pandoc fenced div
syntax is used to print the tables in two columns by sex. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{:::: \{style="display: flex;"\}

::: \{\}

**BV/TV**

|Sex |Genotype |  n| Mean|  SEM|     P|Sig |
|:---|:--------|--:|----:|----:|-----:|:---|
|M   |Cre      |  3| 18.5| 1.58|      |    |
|M   |Cre;fl   |  3| 19.9| 1.11| 0.701|    |


:::

::: \{\}

**BV/TV**

|Sex |Genotype |  n| Mean|   SEM|     P|Sig |
|:---|:--------|--:|----:|-----:|-----:|:---|
|F   |Cre      |  3| 18.3| 1.041|      |    |
|F   |Cre;fl   |  3| 17.1| 0.936| 0.656|    |


:::

::::
}\if{html}{\out{</div>}}
}
\description{
This function is a wraps and extends \code{\link[knitr:kable]{knitr::kable()}} to format and print the
results of bone microCT comparison analyses at one site by sex for use in an
R Markdown document. A code chunk containing this function should probably
have the \code{results = "asis"} option to allow pretty formatting. Any \code{NA}
values are printed as an empty string.
}
\examples{
gen_key <- read_key_csv(mctr_ex("example-gen-key.csv"))
gen_trab <- read_trabecular_csv(mctr_ex("example-trabecular.csv"),
                            gen_key)
Sp.trab <- gen_trab |> dplyr::filter(Site == "Spine") |> compare_genotypes()
print_results(Sp.trab)
}
