% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_key_csv}
\alias{read_key_csv}
\title{Read in a sample definition key file}
\usage{
read_key_csv(file, ...)
}
\arguments{
\item{file}{The file path to the \code{key.csv} file.}

\item{...}{Additional arguments passed on to \code{readr::read_csv()}.}
}
\value{
A tibble/data frame containing the sample definitions.
}
\description{
Read in a sample definition key file
}
\examples{
key <- read_key_csv("data/key.csv")
}
