% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_microCTr}
\alias{update_microCTr}
\title{Update microCTr}
\usage{
update_microCTr()
}
\description{
This is a convenience function wrapper aroung \code{\link[remotes:install_github]{remotes::install_github()}} to
install/update \link{microCTr}. Meant to be used as an RStudio addin.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
update_microCTr()
\dontshow{\}) # examplesIf}
}
